+++ dmn/tovar.php
@@ -64,9 +64,14 @@ if ($_SESSION['auth_admin'] == 'yes_auth') {
             $id = $_GET['id'];
         }
         switch ($action) {
+
             case 'delete':
-                $sth_delete = DB::getStatement('DELETE FROM table_products WHERE products_id = ?');
-                $sth_delete->execute(array($id));
+                if ($_SESSION['delete_tovar'] == 1) {
+                    $sth_delete = DB::getStatement('DELETE FROM table_products WHERE products_id = ?');
+                    $sth_delete->execute(array($id));
+                } else {
+                    $msgerror = 'У вас нет прав на удаление товаров!';
+                }
             break;
         }
     }
@@ -80,10 +85,12 @@ if ($_SESSION['auth_admin'] == 'yes_auth') {
         <link href="css/reset.css" rel="stylesheet" type="text/css" />
         <link href="css/style.css" rel="stylesheet" type="text/css" />
         <link href="js/jquery_confirm/jquery.confirm/jquery.confirm.css" rel="stylesheet" type="text/css" />
+        <link href="js/fancyBox/source/jquery.fancybox.css" rel="stylesheet" type="text/css" />
         <script type="text/javascript" src="js/jquery-2.1.1.js"></script>
         <script type="text/javascript" src="js/jquery.migrate.js"></script>
         <script type="text/javascript" src="js/admin-script.js"></script>
         <script type="text/javascript" src="js/jquery_confirm/jquery.confirm/jquery.confirm.js"></script>
+        <script type="text/javascript" src="js/fancyBox/source/jquery.fancybox.js" ></script>
     </head>
     <body>
     <div id="block-body">
@@ -149,6 +156,7 @@ if ($_SESSION['auth_admin'] == 'yes_auth') {
             </div>
             <ul id="block-tovar">
         <?php
+        if (isset($msgerror)) echo '<p id="form-error" align="center">'.$msgerror.'</p>';
             /**
             * Постраничная навигация
             */
