+++ dmn/view_order.php
@@ -28,13 +28,21 @@ if ($_SESSION['auth_admin'] == 'yes_auth') {
     }
     switch ($action) {
         case 'accept':
-            $sth = DB::getStatement("UPDATE orders SET order_confirmed = 'yes' WHERE order_id = ?");
-            $sth->execute(array($id));
+            if ($_SESSION['accept_orders'] == '1') {
+                $sth = DB::getStatement("UPDATE orders SET order_confirmed = 'yes' WHERE order_id = ?");
+                $sth->execute(array($id));
+            } else {
+                $msgerror = 'У вас нет прав на одобрение заказов!';
+            }
             break;
         case 'delete':
-            $sth = DB::getStatement("DELETE FROM orders WHERE order_id = ?");
-            $sth->execute(array($id));
-            header("Location: orders.php");
+            if ($_SESSION['delete_orders'] == '1') {
+                $sth = DB::getStatement("DELETE FROM orders WHERE order_id = ?");
+                $sth->execute(array($id));
+                header("Location: orders.php");
+            } else {
+                $msgerror = 'У вас нет прав на удаление заказов!';
+            }
             break;
     }
 
@@ -49,10 +57,12 @@ if ($_SESSION['auth_admin'] == 'yes_auth') {
         <link href="css/reset.css" rel="stylesheet" type="text/css" />
         <link href="css/style.css" rel="stylesheet" type="text/css" />
         <link href="js/jquery_confirm/jquery.confirm/jquery.confirm.css" rel="stylesheet" type="text/css" />
+        <link href="js/fancyBox/source/jquery.fancybox.css" rel="stylesheet" type="text/css" />
         <script type="text/javascript" src="js/jquery-2.1.1.js"></script>
         <script type="text/javascript" src="js/jquery.migrate.js"></script>
         <script type="text/javascript" src="js/admin-script.js"></script>
         <script type="text/javascript" src="js/jquery_confirm/jquery.confirm/jquery.confirm.js"></script>
+        <script type="text/javascript" src="js/fancyBox/source/jquery.fancybox.js" ></script>
     </head>
     <body>
     <div id="block-body">
@@ -66,84 +76,89 @@ if ($_SESSION['auth_admin'] == 'yes_auth') {
             </div>
 
             <?php
-            $sth_orders_select = DB::getStatement('SELECT * FROM orders WHERE order_id=?');
-            $sth_orders_select->execute(array($id));
-            $rows_orders = $sth_orders_select->fetchAll();
-            if(!empty($rows_orders)) {
-                foreach ($rows_orders as $row_orders) {
-                    if($row_orders['order_confirmed'] == 'yes') {
-                        $status = '<span class="green">Обработан</span>';
-                    } else {
-                        $status = '<span class="red">Не обработан</span>';
-                    }
-                    echo '
-                    <p class="view-order-link">
-                        <a class="green" href="view_order.php?id='.$row_orders['order_id'].'&action=accept">Подтвердить заказ</a> |
-                        <a class="delete" rel="view_order.php?id='.$row_orders['order_id'].'&action=delete">Удалить заказ</a>
-                    </p>
-                    <p class="order-datetime">'.$row_orders['order_datetime'].'</p>
-                    <p class="order-number">Заказ № '.$row_orders['order_id'].' - '.$status.'</p>
-                    <table align="center" cellpadding="10" width="100%">
-                        <tr>
-                            <th>№</th>
-                            <th>Наименование товара</th>
-                            <th>Цена</th>
-                            <th>Количество</th>
-                        </tr>';
-                    $sth_products = DB::getStatement('SELECT * FROM buy_products, table_products WHERE buy_products.buy_id_order=? AND table_products.products_id=buy_products.buy_id_product');
-                    $sth_products->execute(array($id));
-                    $rows_products = $sth_products->fetchAll();
-                    if (!empty($rows_products)) {
-                        foreach( $rows_products as $row_product) {
-                            $price = $price + $row_product['price'] * $row_product['buy_count_product'];
-                            $index_count = $index_count + 1;
-                            echo '
-                            <tr>
-                                <td align="center">'.$index_count.'</td>
-                                <td align="center">'.$row_product["title"].'</td>
-                                <td align="center">'.$row_product["price"].'</td>
-                                <td align="center">'.$row_product["buy_count_product"].'</td>
-                            </tr>
-                            ';
+            if (isset($msgerror)) echo '<p id="form-error" align="center">'.$msgerror.'</p>';
+            if ($_SESSION['view_orders'] == '1') {
+
+                $sth_orders_select = DB::getStatement('SELECT * FROM orders WHERE order_id=?');
+                $sth_orders_select->execute(array($id));
+                $rows_orders = $sth_orders_select->fetchAll();
+                if(!empty($rows_orders)) {
+                    foreach ($rows_orders as $row_orders) {
+                        if($row_orders['order_confirmed'] == 'yes') {
+                            $status = '<span class="green">Обработан</span>';
+                        } else {
+                            $status = '<span class="red">Не обработан</span>';
                         }
+                        echo '
+                        <p class="view-order-link">
+                            <a class="green" href="view_order.php?id='.$row_orders['order_id'].'&action=accept">Подтвердить заказ</a> |
+                            <a class="delete" rel="view_order.php?id='.$row_orders['order_id'].'&action=delete">Удалить заказ</a>
+                        </p>
+                        <p class="order-datetime">'.$row_orders['order_datetime'].'</p>
+                        <p class="order-number">Заказ № '.$row_orders['order_id'].' - '.$status.'</p>
+                        <table align="center" cellpadding="10" width="100%">
+                            <tr>
+                                <th>№</th>
+                                <th>Наименование товара</th>
+                                <th>Цена</th>
+                                <th>Количество</th>
+                            </tr>';
+                        $sth_products = DB::getStatement('SELECT * FROM buy_products, table_products WHERE buy_products.buy_id_order=? AND table_products.products_id=buy_products.buy_id_product');
+                        $sth_products->execute(array($id));
+                        $rows_products = $sth_products->fetchAll();
+                        if (!empty($rows_products)) {
+                            foreach( $rows_products as $row_product) {
+                                $price = $price + $row_product['price'] * $row_product['buy_count_product'];
+                                $index_count = $index_count + 1;
+                                echo '
+                                <tr>
+                                    <td align="center">'.$index_count.'</td>
+                                    <td align="center">'.$row_product["title"].'</td>
+                                    <td align="center">'.$row_product["price"].'</td>
+                                    <td align="center">'.$row_product["buy_count_product"].'</td>
+                                </tr>
+                                ';
+                            }
+                        }
+
+                        if($row_orders['order_pay'] == 'accepted') {
+                            $statpay = '<span class="green">Оплачено</span>';
+                        } else {
+                            $statpay = '<span class="red">Не оплачено</span>';
+                        }
+
+                        echo '
+                        </table>
+                        <ul id="info-order">
+                            <li>Общая цена - <span>'.$price.'</span></li>
+                            <li>Способ доставки - <span>'.$row_orders["order_dostavka"].'</span></li>
+                            <li>Статус оплаты - '.$statpay.'</li>
+                            <li>Тип оплаты - <span>'.$row_orders["order_type_pay"].'</span></li>
+                            <li>Дата оплаты - <span>'.$row_orders["order_datetime"].'</span></li>
+                        </ul>
+
+                        <table align="center" cellpadding="10" width="100%">
+                            <tr>
+                                <th>ФИО</th>
+                                <th>Адрес</th>
+                                <th>Контакты</th>
+                                <th>Примечание</th>
+                            </tr>
+                            <tr>
+                                <td align="center">'.$row_orders["order_fio"].'</td>
+                                <td align="center">'.$row_orders["order_address"].'</td>
+                                <td align="center">'.$row_orders["order_phone"].'<br />'.$row_orders["order_email"].'</td>
+                                <td align="center">'.$row_orders["order_note"].'</td>
+                            </tr>
+                        </table
+                        ';
+
                     }
-
-                    if($row_orders['order_pay'] == 'accepted') {
-                        $statpay = '<span class="green">Оплачено</span>';
-                    } else {
-                        $statpay = '<span class="red">Не оплачено</span>';
-                    }
-
-                    echo '
-                    </table>
-                    <ul id="info-order">
-                        <li>Общая цена - <span>'.$price.'</span></li>
-                        <li>Способ доставки - <span>'.$row_orders["order_dostavka"].'</span></li>
-                        <li>Статус оплаты - '.$statpay.'</li>
-                        <li>Тип оплаты - <span>'.$row_orders["order_type_pay"].'</span></li>
-                        <li>Дата оплаты - <span>'.$row_orders["order_datetime"].'</span></li>
-                    </ul>
-
-                    <table align="center" cellpadding="10" width="100%">
-                        <tr>
-                            <th>ФИО</th>
-                            <th>Адрес</th>
-                            <th>Контакты</th>
-                            <th>Примечание</th>
-                        </tr>
-                        <tr>
-                            <td align="center">'.$row_orders["order_fio"].'</td>
-                            <td align="center">'.$row_orders["order_address"].'</td>
-                            <td align="center">'.$row_orders["order_phone"].'<br />'.$row_orders["order_email"].'</td>
-                            <td align="center">'.$row_orders["order_note"].'</td>
-                        </tr>
-                    </table
-                    ';
 
                 }
-
+            } else {
+                echo '<p id="form-error" align="center">У вас нет прав на просмотр данного раздела!</p>';
             }
-
             ?>
         </div>
     </div>
